python_engine/requirements.txt
annotated-types
anyio
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
deprecation
distro
fsspec
google-auth
google-genai
h11
h2
hpack
httpcore
httpx
hyperframe
idna
markdown-it-py
mdurl
mmh3
multidict
packaging
postgrest
propcache
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
Pygments
pyiceberg
PyJWT
pyparsing
pyroaring
python-dateutil
python-dotenv
realtime
requests
rich
six
sniffio
storage3
StrEnum
strictyaml
supabase
supabase-auth
supabase-functions
tenacity
typing-inspection
typing_extensions
urllib3
websockets
yarl
zstandard